# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A high-performance subscription tracking service built with Go and PostgreSQL.\nThis API handles user authentication, subscription lifecycle management, and expense analytics.",
        "title": "SubScout Management API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Ayan Chatterjee",
            "url": "https://github.com/Ayan25844",
            "email": "ayan25844@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/categories": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "User route to get a list of all categories",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Routes"
                ],
                "summary": "Get the list of all categories",
                "responses": {
                    "200": {
                        "description": "List of categories retrieved successfully",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Category"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized: Missing token",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "No categories found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Admin route to create a unique subscription category",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Routes"
                ],
                "summary": "Create a new category",
                "parameters": [
                    {
                        "description": "Category Name",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Category_DTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Category created successfully",
                        "schema": {
                            "$ref": "#/definitions/Category"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized: Missing token",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden - Insufficient Permissions",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Category already exists",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/categories/{id}": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Admin route to rename a category",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Routes"
                ],
                "summary": "Update the name of a category",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Renamed category",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Category_DTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category updated successfully",
                        "schema": {
                            "$ref": "#/definitions/Category"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized: Missing token",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden: Insufficient Permissions",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Category not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Category already exists",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/subscriptions": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "User route to get a list of all subscriptions for the logged in user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Routes"
                ],
                "summary": "Get a list of all subscriptions",
                "responses": {
                    "200": {
                        "description": "List of subscriptions retrieved successfully",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Subscription"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized: Missing token",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "No subscriptions found for this user",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "User route to create a unique subscription for the logged in user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Routes"
                ],
                "summary": "Create a new subscription",
                "parameters": [
                    {
                        "description": "New subscription",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Subscription_DTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Subscription created successfully",
                        "schema": {
                            "$ref": "#/definitions/Subscription"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized: Missing token",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Subscription already exists",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Category": {
            "description": "Metadata used to organize and filter subscriptions",
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2026-03-09T12:00:00Z"
                },
                "created_by": {
                    "type": "string",
                    "format": "uuid"
                },
                "id": {
                    "type": "string",
                    "format": "uuid"
                },
                "name": {
                    "type": "string",
                    "example": "Cloud"
                },
                "updated_at": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2026-03-09T12:00:00Z"
                },
                "updated_by": {
                    "type": "string",
                    "format": "uuid"
                }
            }
        },
        "Category_DTO": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Cloud"
                }
            }
        },
...
        "Subscription_DTO": {
            "type": "object",
            "required": [
                "category_id",
                "monthly_cost",
                "next_billing_date",
                "service_name"
            ],
            "properties": {
                "billing_cycle": {
                    "type": "string",
                    "example": "monthly"
                },
                "category_id": {
                    "type": "string",
                    "format": "uuid"
                },
                "monthly_cost": {
                    "type": "number",
                    "format": "float",
                    "example": 15.99
                },
                "next_billing_date": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2026-03-09T12:00:00Z"
                },
                "service_name": {
                    "type": "string",
                    "example": "Catalog service"
                },
                "trial_end_date": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2026-03-09T12:00:00Z"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "JWT Authorization header using the Bearer scheme.\nExample: \"Authorization: Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}